Refactor and Enhance Game Functionality
This commit includes several improvements and refactorings to the Alien Invasion game:

1. Ship Movement: Improved the ship's movement logic to ensure smoother and more responsive controls.
2. Bullet Management: Enhanced bullet management to handle collisions and screen boundaries more efficiently.
3. Alien Fleet: Refactored the alien fleet creation and movement logic for better performance and readability.
4. Scoreboard: Updated the scoreboard to dynamically display the current score, high score, and remaining ships.
5. Game Initialization: Improved game initialization and reset procedures to ensure a seamless gaming experience.
6. Button Handling: Enhanced button handling to provide better user interaction and feedback.
7. Code Cleanup: Removed redundant code and improved overall code readability and maintainability.
