+++ ship.py
@@ -1,0 +1,43 @@
+import pygame
+from pygame.sprite import Sprite
+
+class Ship(Sprite):
+    """Class to manage the player's ship."""
+
+    def __init__(self, game):
+        """Initialize ship and set starting position."""
+        super().__init__()
+        self.screen = game.screen
+        self.settings = game.settings
+        self.screen_rect = game.screen.get_rect()
+
+        # Load ship image and initialize dimensions
+        self.image = pygame.image.load("images/spaceship.bmp")
+        self.rect = self.image.get_rect()
+        self.rect.midbottom = self.screen_rect.midbottom
+
+        # Store decimal position
+        self.x = float(self.rect.x)
+
+        # Movement flags
+        self.moving_right = False
+        self.moving_left = False
+
+    def update(self):
+        """Update the ship's position based on movement flag."""
+        if self.moving_right and self.rect.right < self.screen_rect.right:
+            self.x += self.settings.ship_speed
+        if self.moving_left and self.rect.left > 0:
+            self.x -= self.settings.ship_speed
+
+        # Update ret object from self.x
+        self.rect.x = self.x
+
+    def blitme(self):
+        """Draw the ship at its current location."""
+        self.screen.blit(self.image, self.rect)
+
+    def center_ship(self):
+        """Center the ship on the screen."""
+        self.rect.midbottom = self.screen_rect.midbottom
+        self.x = float(self.rect.x)
